# REFACTOR - ocsp.js
- pkijs 관련 함수들은 따로 분리함
  * 다른 곳에서 사용할 가능성이 있음
  * 여러 보안 라이브러리가 혼합해서 사용되다 보니까 나중에 헷갈릴 수가 있음
